requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Templates
jinja2>=3.1.3

# Validation & Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# QR Code for packing labels
qrcode[pil]>=7.4.2

# PDF Generation & Merging
pypdf>=3.17.0

# Platform Integrations
httpx>=0.24.0
apscheduler>=3.10.0
schedule>=1.2.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4